+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## Changed
+
+- 💄 title bar is now the same color as the side panel on Windows 11
+
 ## [v5.1.0] - 2026-02-06
 
 ### Added